Renamed prim/Tuple.scala to prim/tuple.scala Made requested changes
