+++ roscala/src/main/scala/coop/rchain/rosette/prim/tuple.scala
@@ -1,0 +1,28 @@
+package coop.rchain.rosette.prim
+
+import coop.rchain.rosette.{Ctxt, Ob, RblBool, Tuple}
+import coop.rchain.rosette.macros.{checkArgumentMismatch, checkTypeMismatch}
+import coop.rchain.rosette.prim.Prim._
+
+object tuple {
+  object tplCons extends Prim {
+    override val name: String = "tuple-cons"
+    override val minArgs: Int = 2
+    override val maxArgs: Int = 2
+
+    @checkArgumentMismatch
+    override def fn(ctxt: Ctxt): Either[PrimError, Tuple] = {
+      val elem = ctxt.argvec.elem
+      checkTuple(1, elem).map(Tuple.cons(elem.head, _)) // arg1 must be a Tuple
+    }
+
+  }
+
+  private def checkTuple(n: Int, elem: Seq[Ob]): Either[PrimError, Tuple] =
+    if (!elem(n).isInstanceOf[Tuple]) {
+      Left(TypeMismatch(n, Tuple.getClass().getName()))
+    } else {
+      Right(elem(n).asInstanceOf[Tuple])
+    }
+
+}
